music-finder-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
pydantic==2.4.2
python-multipart==0.0.6 